refactor range slider logic
